changes based on pr comments
