+++ services/common/src/components/common/LinkButton.js
@@ -7,7 +7,7 @@ const propTypes = {
   children: PropTypes.any,
   disabled: PropTypes.bool,
   title: PropTypes.string,
-  class: PropTypes.string,
+  className: PropTypes.string,
 };
 
 const defaultProps = {
@@ -23,7 +23,7 @@ const LinkButton = (props) => (
     onKeyPress={props.onClick}
     tabIndex={props.tabIndex}
     disabled={props.disabled}
-    class={props.class}
+    className={props.className}
     {...props}
   >
     {props.children}
